fix&add: RankingDto

+++ src/main/java/klab/sugangstar/dto/characterDTO/CharacterCreateDto.java
@@ -8,12 +8,12 @@ import java.util.List;
 public class CharacterCreateDto {
     private Long userId;
     private String nickName;
-    private int memorization;
-    private int concentration;
-    private int patience;
-    private int creativity;
-    private int metacognition;
-    private int understanding;
+    private float memorization;
+    private float concentration;
+    private float patience;
+    private float creativity;
+    private float metacognition;
+    private float understanding;
     private int semester;
     private List<Long> subjectIds;
 }
