+++ src/main/java/klab/sugangstar/dto/characterDTO/CharacterCreateDto3.java
@@ -7,11 +7,11 @@ import java.util.List;
 @Data
 public class CharacterCreateDto3 {
     private Long characterId;
-    private int memorization;
-    private int concentration;
-    private int patience;
-    private int creativity;
-    private int metacognition;
-    private int understanding;
+    private float memorization;
+    private float concentration;
+    private float patience;
+    private float creativity;
+    private float metacognition;
+    private float understanding;
     private List<Long> subjectIds;
 }
